feat: 修改 long long 位 int64_t

+++ imlibV2/src/main/cpp/rc_adapter.cpp
@@ -11,7 +11,7 @@
 #include "jni_header/NativeStringCallback_jni.h"
 #include "jni_header/NativeIntListener_jni.h"
 
-int create_engine_builder(RcimEngineBuilderParam *param, long long *outBuilderPtr) {
+int create_engine_builder(RcimEngineBuilderParam *param, int64_t *outBuilderPtr) {
     if (!param) {
         return 0;
     }
@@ -20,21 +20,21 @@ int create_engine_builder(RcimEngineBuilderParam *param, long long *outBuilderPt
     RcimEngineBuilder *builder;
     RcimEngineError code = rcim_create_engine_builder(param, &builder);
     // 向输出参数添加数据
-    *outBuilderPtr = static_cast<long long>(reinterpret_cast<uintptr_t>(builder));
+    *outBuilderPtr = static_cast<int64_t>(reinterpret_cast<uintptr_t>(builder));
     return code;
 }
 
-long long rcim_sdk_version_array_new(int size) {
+int64_t rcim_sdk_version_array_new(int size) {
     RcimSDKVersion *array = (RcimSDKVersion *) malloc(sizeof(RcimSDKVersion) * size);
     if (!array) return 0; // 内存分配失败
-    return (long long) (uintptr_t) array; // 转成 long long 类型安全返回
+    return (int64_t) (uintptr_t) array; // 转成 int64_t 类型安全返回
 }
 
-void rcim_sdk_version_array_insert(long long ptr, long *ptrArr, int size) {
+void rcim_sdk_version_array_insert(int64_t ptr, int64_t *ptrArr, int size) {
     if (ptr == 0) return;
     RcimSDKVersion *array = (RcimSDKVersion *) ptr;
     for (size_t i = 0; i < size; ++i) {
-        long ptr = ptrArr[i];
+        int64_t ptr = ptrArr[i];
         RcimSDKVersion *ver = (RcimSDKVersion *) ptr;
         array[i].name = ver->name;
         array[i].version = ver->version;
@@ -42,13 +42,13 @@ void rcim_sdk_version_array_insert(long long ptr, long *ptrArr, int size) {
 }
 
 
-void rcim_sdk_version_array_free(long long ptr) {
+void rcim_sdk_version_array_free(int64_t ptr) {
     if (ptr == 0) return;
     RcimSDKVersion *array = (RcimSDKVersion *) (intptr_t) ptr;
     free(array);
 }
 
-int engine_builder_set_store_path(long long builderPtr, const char *storePath) {
+int engine_builder_set_store_path(int64_t builderPtr, const char *storePath) {
     auto *builder = reinterpret_cast<RcimEngineBuilder *>(static_cast<uintptr_t>(builderPtr));
 
     RcimEngineError code = rcim_engine_builder_set_store_path(builder, storePath);
@@ -91,13 +91,13 @@ int engine_builder_set_store_path(long long builderPtr, const char *storePath) {
     return code;
 }
 
-int engine_builder_build(long long builderPtr, long long *outEnginePtr) {
+int engine_builder_build(int64_t builderPtr, int64_t *outEnginePtr) {
     auto *builder = reinterpret_cast<RcimEngineBuilder *>(static_cast<uintptr_t>(builderPtr));
     RcimEngineSync *engine_ = nullptr;
     RcimEngineError code = rcim_engine_builder_build(builder, &engine_);
 
     // 向输出参数添加数据
-    *outEnginePtr = static_cast<long long>(reinterpret_cast<uintptr_t>(engine_));
+    *outEnginePtr = static_cast<int64_t>(reinterpret_cast<uintptr_t>(engine_));
     return code;
 }
 
@@ -107,7 +107,7 @@ void engine_connect_adapter(const void *context, enum RcimEngineError code, cons
     releaseContextByCallback(context);
 }
 
-void engine_connect(long long enginePtr, const char *token, int timeout, void *callback) {
+void engine_connect(int64_t enginePtr, const char *token, int timeout, void *callback) {
     auto *engine = reinterpret_cast<RcimEngineSync *>(static_cast<uintptr_t>(enginePtr));
     void *context = genContextByCallback(engine, callback);
     rcim_engine_connect(engine, token, timeout, context, engine_connect_adapter);
@@ -118,7 +118,7 @@ void engine_set_connection_status_listener_adapter(const void *context,
     rcim::callNativeIntListener(context, status);
 }
 
-void engine_set_connection_status_listener(long long enginePtr, void *listener) {
+void engine_set_connection_status_listener(int64_t enginePtr, void *listener) {
     auto *engine = reinterpret_cast<RcimEngineSync *>(static_cast<uintptr_t>(enginePtr));
     void *context = genContextByListener(engine, listener);
     rcim_engine_set_connection_status_listener(engine, context,
@@ -135,7 +135,7 @@ void engine_send_message_adapter(const void *context,
     int a = 2;
 }
 
-void engine_send_message(long long enginePtr, RcimMessageBox *msgBox, void *sendMsgCallback) {
+void engine_send_message(int64_t enginePtr, RcimMessageBox *msgBox, void *sendMsgCallback) {
     // todo qixinbing
     auto *engine = reinterpret_cast<RcimEngineSync *>(static_cast<uintptr_t>(enginePtr));
     rcim_engine_send_message(engine, msgBox, nullptr, nullptr, engine_send_message_adapter,
